pyproject.toml
[tool.poetry]
name = "cmc_bot_tst"
version = "0.1.0"
description = ""
authors = ["erastoff <yuriy.erastov@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.8.1"
requests = "^2.32.3"
python-telegram-bot = "^21.2"
aiogram = "^3.7.0"
loguru = "^0.7.2"
aiohttp = "^3.9.5"
aiosqlite = "^0.20.0"
pydantic-settings = "^2.3.3"
python-decouple = "^3.8"

[tool.poetry.dev-dependencies]
black = "^24.4.2"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
